+++ sentry-android-replay/src/main/java/io/sentry/android/replay/capture/BaseCaptureStrategy.kt
@@ -1,45 +1,55 @@
 package io.sentry.android.replay.capture
 
 import android.view.MotionEvent
+import io.sentry.Breadcrumb
 import io.sentry.DateUtils
-import io.sentry.Hint
 import io.sentry.IHub
-import io.sentry.ReplayRecording
 import io.sentry.SentryOptions
-import io.sentry.SentryReplayEvent
 import io.sentry.SentryReplayEvent.ReplayType
+import io.sentry.SentryReplayEvent.ReplayType.BUFFER
 import io.sentry.SentryReplayEvent.ReplayType.SESSION
 import io.sentry.android.replay.ReplayCache
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_BIT_RATE
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_FRAME_RATE
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_HEIGHT
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_ID
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_REPLAY_ID
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_REPLAY_RECORDING
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_REPLAY_SCREEN_AT_START
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_REPLAY_TYPE
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_TIMESTAMP
+import io.sentry.android.replay.ReplayCache.Companion.SEGMENT_KEY_WIDTH
 import io.sentry.android.replay.ScreenshotRecorderConfig
+import io.sentry.android.replay.capture.CaptureStrategy.Companion.createSegment
+import io.sentry.android.replay.capture.CaptureStrategy.Companion.currentEventsLock
+import io.sentry.android.replay.capture.CaptureStrategy.ReplaySegment
+import io.sentry.android.replay.util.PersistableLinkedList
 import io.sentry.android.replay.util.gracefullyShutdown
 import io.sentry.android.replay.util.submitSafely
 import io.sentry.protocol.SentryId
-import io.sentry.rrweb.RRWebBreadcrumbEvent
 import io.sentry.rrweb.RRWebEvent
 import io.sentry.rrweb.RRWebIncrementalSnapshotEvent
 import io.sentry.rrweb.RRWebInteractionEvent
 import io.sentry.rrweb.RRWebInteractionEvent.InteractionType
 import io.sentry.rrweb.RRWebInteractionMoveEvent
 import io.sentry.rrweb.RRWebInteractionMoveEvent.Position
-import io.sentry.rrweb.RRWebMetaEvent
-import io.sentry.rrweb.RRWebVideoEvent
 import io.sentry.transport.ICurrentDateProvider
-import io.sentry.util.FileUtils
 import java.io.File
 import java.util.Date
 import java.util.LinkedList
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.ThreadFactory
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicLong
 import java.util.concurrent.atomic.AtomicReference
+import kotlin.properties.ReadWriteProperty
+import kotlin.reflect.KProperty
 
 internal abstract class BaseCaptureStrategy(
     private val options: SentryOptions,
     private val hub: IHub?,
     private val dateProvider: ICurrentDateProvider,
-    protected var recorderConfig: ScreenshotRecorderConfig,
     executor: ScheduledExecutorService? = null,
     private val replayCacheProvider: ((replayId: SentryId, recorderConfig: ScreenshotRecorderConfig) -> ReplayCache)? = null
 ) : CaptureStrategy {
@@ -52,16 +62,38 @@ internal abstract class BaseCaptureStrategy(
         private const val CAPTURE_MOVE_EVENT_THRESHOLD = 500
     }
 
+    private val persistingExecutor: ScheduledExecutorService by lazy {
+        Executors.newSingleThreadScheduledExecutor(ReplayPersistingExecutorServiceThreadFactory())
+    }
+
+    protected val isTerminating = AtomicBoolean(false)
     protected var cache: ReplayCache? = null
-    protected val segmentTimestamp = AtomicReference<Date>()
+    protected var recorderConfig: ScreenshotRecorderConfig by persistableAtomic { _, _, newValue ->
+        if (newValue == null) {
+            // recorderConfig is only nullable on init, but never after
+            return@persistableAtomic
+        }
+        cache?.persistSegmentValues(SEGMENT_KEY_HEIGHT, newValue.recordingHeight.toString())
+        cache?.persistSegmentValues(SEGMENT_KEY_WIDTH, newValue.recordingWidth.toString())
+        cache?.persistSegmentValues(SEGMENT_KEY_FRAME_RATE, newValue.frameRate.toString())
+        cache?.persistSegmentValues(SEGMENT_KEY_BIT_RATE, newValue.bitRate.toString())
+    }
+    protected var segmentTimestamp by persistableAtomicNullable<Date>(propertyName = SEGMENT_KEY_TIMESTAMP) { _, _, newValue ->
+        cache?.persistSegmentValues(SEGMENT_KEY_TIMESTAMP, if (newValue == null) null else DateUtils.getTimestamp(newValue))
+    }
     protected val replayStartTimestamp = AtomicLong()
-    protected val screenAtStart = AtomicReference<String>()
-    override val currentReplayId = AtomicReference(SentryId.EMPTY_ID)
-    override val currentSegment = AtomicInteger(0)
+    protected var screenAtStart by persistableAtomicNullable<String>(propertyName = SEGMENT_KEY_REPLAY_SCREEN_AT_START)
+    override var currentReplayId: SentryId by persistableAtomic(initialValue = SentryId.EMPTY_ID, propertyName = SEGMENT_KEY_REPLAY_ID)
+    override var currentSegment: Int by persistableAtomic(initialValue = -1, propertyName = SEGMENT_KEY_ID)
     override val replayCacheDir: File? get() = cache?.replayCacheDir
 
-    protected val currentEvents = LinkedList<RRWebEvent>()
-    private val currentEventsLock = Any()
+    private var replayType by persistableAtomic<ReplayType>(propertyName = SEGMENT_KEY_REPLAY_TYPE)
+    protected val currentEvents: LinkedList<RRWebEvent> = PersistableLinkedList(
+        propertyName = SEGMENT_KEY_REPLAY_RECORDING,
+        options,
+        persistingExecutor,
+        cacheProvider = { cache }
+    )
     private val currentPositions = LinkedHashMap<Int, ArrayList<Position>>(10)
     private var touchMoveBaseline = 0L
     private var lastCapturedMoveEvent = 0L
@@ -70,170 +102,67 @@ internal abstract class BaseCaptureStrategy(
         executor ?: Executors.newSingleThreadScheduledExecutor(ReplayExecutorServiceThreadFactory())
     }
 
-    override fun start(segmentId: Int, replayId: SentryId, cleanupOldReplays: Boolean) {
-        currentSegment.set(segmentId)
-        currentReplayId.set(replayId)
+    override fun start(
+        recorderConfig: ScreenshotRecorderConfig,
+        segmentId: Int,
+        replayId: SentryId
+    ) {
+        cache = replayCacheProvider?.invoke(replayId, recorderConfig) ?: ReplayCache(options, replayId, recorderConfig)
 
-        if (cleanupOldReplays) {
-            replayExecutor.submitSafely(options, "$TAG.replays_cleanup") {
-                // clean up old replays
-                options.cacheDirPath?.let { cacheDir ->
-                    File(cacheDir).listFiles { dir, name ->
-                        // TODO: also exclude persisted replay_id from scope when implementing ANRs
-                        if (name.startsWith("replay_") && !name.contains(
-                                currentReplayId.get().toString()
-                            )
-                        ) {
-                            FileUtils.deleteRecursively(File(dir, name))
-                        }
-                        false
-                    }
-                }
-            }
-        }
+        // TODO: this should be persisted even after conversion
+        replayType = if (this is SessionCaptureStrategy) SESSION else BUFFER
+        this.recorderConfig = recorderConfig
+        currentSegment = segmentId
+        currentReplayId = replayId
 
-        cache =
-            replayCacheProvider?.invoke(replayId, recorderConfig) ?: ReplayCache(options, replayId, recorderConfig)
-
-        // TODO: replace it with dateProvider.currentTimeMillis to also test it
-        segmentTimestamp.set(DateUtils.getCurrentDateTime())
+        segmentTimestamp = DateUtils.getCurrentDateTime()
         replayStartTimestamp.set(dateProvider.currentTimeMillis)
-        // TODO: finalize old recording if there's some left on disk and send it using the replayId from persisted scope (e.g. for ANRs)
     }
 
     override fun resume() {
-        // TODO: replace it with dateProvider.currentTimeMillis to also test it
-        segmentTimestamp.set(DateUtils.getCurrentDateTime())
+        segmentTimestamp = DateUtils.getCurrentDateTime()
     }
 
     override fun pause() = Unit
 
     override fun stop() {
         cache?.close()
-        currentSegment.set(0)
+        currentSegment = -1
         replayStartTimestamp.set(0)
-        segmentTimestamp.set(null)
-        currentReplayId.set(SentryId.EMPTY_ID)
+        segmentTimestamp = null
+        currentReplayId = SentryId.EMPTY_ID
     }
 
-    protected fun createSegment(
+    protected fun createSegmentInternal(
         duration: Long,
         currentSegmentTimestamp: Date,
         replayId: SentryId,
         segmentId: Int,
         height: Int,
         width: Int,
-        replayType: ReplayType = SESSION
-    ): ReplaySegment {
-        val generatedVideo = cache?.createVideoOf(
+        replayType: ReplayType = SESSION,
+        cache: ReplayCache? = this.cache,
+        frameRate: Int = recorderConfig.frameRate,
+        screenAtStart: String? = this.screenAtStart,
+        breadcrumbs: List<Breadcrumb>? = null,
+        events: LinkedList<RRWebEvent> = this.currentEvents
+    ): ReplaySegment =
+        createSegment(
+            hub,
+            options,
             duration,
-            currentSegmentTimestamp.time,
-            segmentId,
-            height,
-            width
-        ) ?: return ReplaySegment.Failed
-
-        val (video, frameCount, videoDuration) = generatedVideo
-        return buildReplay(
-            video,
-            replayId,
             currentSegmentTimestamp,
+            replayId,
             segmentId,
             height,
             width,
-            frameCount,
-            videoDuration,
-            replayType
+            replayType,
+            cache,
+            frameRate,
+            screenAtStart,
+            breadcrumbs,
+            events
         )
-    }
-
-    private fun buildReplay(
-        video: File,
-        currentReplayId: SentryId,
-        segmentTimestamp: Date,
-        segmentId: Int,
-        height: Int,
-        width: Int,
-        frameCount: Int,
-        duration: Long,
-        replayType: ReplayType
-    ): ReplaySegment {
-        val endTimestamp = DateUtils.getDateTime(segmentTimestamp.time + duration)
-        val replay = SentryReplayEvent().apply {
-            eventId = currentReplayId
-            replayId = currentReplayId
-            this.segmentId = segmentId
-            this.timestamp = endTimestamp
-            replayStartTimestamp = segmentTimestamp
-            this.replayType = replayType
-            videoFile = video
-        }
-
-        val recordingPayload = mutableListOf<RRWebEvent>()
-        recordingPayload += RRWebMetaEvent().apply {
-            this.timestamp = segmentTimestamp.time
-            this.height = height
-            this.width = width
-        }
-        recordingPayload += RRWebVideoEvent().apply {
-            this.timestamp = segmentTimestamp.time
-            this.segmentId = segmentId
-            this.durationMs = duration
-            this.frameCount = frameCount
-            size = video.length()
-            frameRate = recorderConfig.frameRate
-            this.height = height
-            this.width = width
-            // TODO: support non-fullscreen windows later
-            left = 0
-            top = 0
-        }
-
-        val urls = LinkedList<String>()
-        hub?.configureScope { scope ->
-            scope.breadcrumbs.forEach { breadcrumb ->
-                if (breadcrumb.timestamp.time >= segmentTimestamp.time &&
-                    breadcrumb.timestamp.time < endTimestamp.time
-                ) {
-                    val rrwebEvent = options
-                        .replayController
-                        .breadcrumbConverter
-                        .convert(breadcrumb)
-
-                    if (rrwebEvent != null) {
-                        recordingPayload += rrwebEvent
-
-                        // fill in the urls array from navigation breadcrumbs
-                        if ((rrwebEvent as? RRWebBreadcrumbEvent)?.category == "navigation") {
-                            urls.add(rrwebEvent.data!!["to"] as String)
-                        }
-                    }
-                }
-            }
-        }
-
-        if (screenAtStart.get() != null && urls.firstOrNull() != screenAtStart.get()) {
-            urls.addFirst(screenAtStart.get())
-        }
-
-        rotateCurrentEvents(endTimestamp.time) { event ->
-            if (event.timestamp >= segmentTimestamp.time) {
-                recordingPayload += event
-            }
-        }
-
-        val recording = ReplayRecording().apply {
-            this.segmentId = segmentId
-            payload = recordingPayload.sortedBy { it.timestamp }
-        }
-
-        replay.urls = urls
-        return ReplaySegment.Created(
-            videoDuration = duration,
-            replay = replay,
-            recording = recording
-        )
-    }
 
     override fun onConfigurationChanged(recorderConfig: ScreenshotRecorderConfig) {
         this.recorderConfig = recorderConfig
@@ -252,20 +181,6 @@ internal abstract class BaseCaptureStrategy(
         replayExecutor.gracefullyShutdown(options)
     }
 
-    protected fun rotateCurrentEvents(
-        until: Long,
-        callback: ((RRWebEvent) -> Unit)? = null
-    ) {
-        synchronized(currentEventsLock) {
-            var event = currentEvents.peek()
-            while (event != null && event.timestamp < until) {
-                callback?.invoke(event)
-                currentEvents.remove()
-                event = currentEvents.peek()
-            }
-        }
-    }
-
     private class ReplayExecutorServiceThreadFactory : ThreadFactory {
         private var cnt = 0
         override fun newThread(r: Runnable): Thread {
@@ -275,25 +190,12 @@ internal abstract class BaseCaptureStrategy(
         }
     }
 
-    protected sealed class ReplaySegment {
-        object Failed : ReplaySegment()
-        data class Created(
-            val videoDuration: Long,
-            val replay: SentryReplayEvent,
-            val recording: ReplayRecording
-        ) : ReplaySegment() {
-            fun capture(hub: IHub?, hint: Hint = Hint()) {
-                hub?.captureReplay(replay, hint.apply { replayRecording = recording })
-            }
-
-            fun setSegmentId(segmentId: Int) {
-                replay.segmentId = segmentId
-                recording.payload?.forEach {
-                    when (it) {
-                        is RRWebVideoEvent -> it.segmentId = segmentId
-                    }
-                }
-            }
+    private class ReplayPersistingExecutorServiceThreadFactory : ThreadFactory {
+        private var cnt = 0
+        override fun newThread(r: Runnable): Thread {
+            val ret = Thread(r, "SentryReplayPersister-" + cnt++)
+            ret.setDaemon(true)
+            return ret
         }
     }
 
@@ -416,4 +318,52 @@ internal abstract class BaseCaptureStrategy(
             else -> null
         }
     }
+
+    private inline fun <T> persistableAtomicNullable(
+        initialValue: T? = null,
+        propertyName: String,
+        crossinline onChange: (propertyName: String?, oldValue: T?, newValue: T?) -> Unit = { _, _, newValue ->
+            cache?.persistSegmentValues(propertyName, newValue.toString())
+        }
+    ): ReadWriteProperty<Any?, T?> =
+        object : ReadWriteProperty<Any?, T?> {
+            private val value = AtomicReference(initialValue)
+
+            private fun runInBackground(task: () -> Unit) {
+                if (options.mainThreadChecker.isMainThread) {
+                    persistingExecutor.submitSafely(options, "$TAG.runInBackground") {
+                        task()
+                    }
+                } else {
+                    task()
+                }
+            }
+
+            init {
+                runInBackground { onChange(propertyName, initialValue, initialValue) }
+            }
+
+            override fun getValue(thisRef: Any?, property: KProperty<*>): T? = value.get()
+
+            override fun setValue(thisRef: Any?, property: KProperty<*>, value: T?) {
+                val oldValue = this.value.getAndSet(value)
+                if (oldValue != value) {
+                    runInBackground { onChange(propertyName, oldValue, value) }
+                }
+            }
+        }
+
+    private inline fun <T> persistableAtomic(
+        initialValue: T? = null,
+        propertyName: String,
+        crossinline onChange: (propertyName: String?, oldValue: T?, newValue: T?) -> Unit = { _, _, newValue ->
+            cache?.persistSegmentValues(propertyName, newValue.toString())
+        }
+    ): ReadWriteProperty<Any?, T> =
+        persistableAtomicNullable<T>(initialValue, propertyName, onChange) as ReadWriteProperty<Any?, T>
+
+    private inline fun <T> persistableAtomic(
+        crossinline onChange: (propertyName: String?, oldValue: T?, newValue: T?) -> Unit
+    ): ReadWriteProperty<Any?, T> =
+        persistableAtomicNullable<T>(null, "", onChange) as ReadWriteProperty<Any?, T>
 }