+++ sentry-android-replay/src/main/java/io/sentry/android/replay/capture/SessionCaptureStrategy.kt
@@ -2,7 +2,6 @@ package io.sentry.android.replay.capture
 
 import android.graphics.Bitmap
 import io.sentry.DateUtils
-import io.sentry.Hint
 import io.sentry.IConnectionStatusProvider.ConnectionStatus.DISCONNECTED
 import io.sentry.IHub
 import io.sentry.SentryLevel.DEBUG
@@ -10,6 +9,7 @@ import io.sentry.SentryLevel.INFO
 import io.sentry.SentryOptions
 import io.sentry.android.replay.ReplayCache
 import io.sentry.android.replay.ScreenshotRecorderConfig
+import io.sentry.android.replay.capture.CaptureStrategy.ReplaySegment
 import io.sentry.android.replay.util.submitSafely
 import io.sentry.protocol.SentryId
 import io.sentry.transport.ICurrentDateProvider
@@ -20,22 +20,25 @@ internal class SessionCaptureStrategy(
     private val options: SentryOptions,
     private val hub: IHub?,
     private val dateProvider: ICurrentDateProvider,
-    recorderConfig: ScreenshotRecorderConfig,
     executor: ScheduledExecutorService? = null,
     replayCacheProvider: ((replayId: SentryId, recorderConfig: ScreenshotRecorderConfig) -> ReplayCache)? = null
-) : BaseCaptureStrategy(options, hub, dateProvider, recorderConfig, executor, replayCacheProvider) {
+) : BaseCaptureStrategy(options, hub, dateProvider, executor, replayCacheProvider) {
 
     internal companion object {
         private const val TAG = "SessionCaptureStrategy"
     }
 
-    override fun start(segmentId: Int, replayId: SentryId, cleanupOldReplays: Boolean) {
-        super.start(segmentId, replayId, cleanupOldReplays)
+    override fun start(
+        recorderConfig: ScreenshotRecorderConfig,
+        segmentId: Int,
+        replayId: SentryId
+    ) {
+        super.start(recorderConfig, segmentId, replayId)
         // only set replayId on the scope if it's a full session, otherwise all events will be
         // tagged with the replay that might never be sent when we're recording in buffer mode
         hub?.configureScope {
-            it.replayId = currentReplayId.get()
-            screenAtStart.set(it.screen)
+            it.replayId = currentReplayId
+            screenAtStart = it.screen
         }
     }
 
@@ -44,7 +47,7 @@ internal class SessionCaptureStrategy(
             if (segment is ReplaySegment.Created) {
                 segment.capture(hub)
 
-                currentSegment.getAndIncrement()
+                currentSegment++
             }
         }
         super.pause()
@@ -62,17 +65,9 @@ internal class SessionCaptureStrategy(
         super.stop()
     }
 
-    override fun sendReplayForEvent(isCrashed: Boolean, eventId: String?, hint: Hint?, onSegmentSent: () -> Unit) {
-        if (!isCrashed) {
-            options.logger.log(DEBUG, "Replay is already running in 'session' mode, not capturing for event %s", eventId)
-        } else {
-            options.logger.log(DEBUG, "Replay is already running in 'session' mode, capturing last segment for crashed event %s", eventId)
-            createCurrentSegment("send_replay_for_event") { segment ->
-                if (segment is ReplaySegment.Created) {
-                    segment.capture(hub, hint ?: Hint())
-                }
-            }
-        }
+    override fun captureReplay(isTerminating: Boolean, onSegmentSent: () -> Unit) {
+        options.logger.log(DEBUG, "Replay is already running in 'session' mode, not capturing for event")
+        this.isTerminating.set(isTerminating)
     }
 
     override fun onScreenshotRecorded(bitmap: Bitmap?, store: ReplayCache.(frameTimestamp: Long) -> Unit) {
@@ -89,43 +84,52 @@ internal class SessionCaptureStrategy(
         replayExecutor.submitSafely(options, "$TAG.add_frame") {
             cache?.store(frameTimestamp)
 
-            val now = dateProvider.currentTimeMillis
-            if ((now - segmentTimestamp.get().time >= options.experimental.sessionReplay.sessionSegmentDuration)) {
-                val currentSegmentTimestamp = segmentTimestamp.get()
-                val segmentId = currentSegment.get()
-                val replayId = currentReplayId.get()
+            val currentSegmentTimestamp = segmentTimestamp
+            currentSegmentTimestamp ?: run {
+                options.logger.log(DEBUG, "Segment timestamp is not set, not recording frame")
+                return@submitSafely
+            }
 
+            if (isTerminating.get()) {
+                options.logger.log(DEBUG, "Not capturing segment, because the app is terminating, will be captured on next launch")
+                return@submitSafely
+            }
+
+            val now = dateProvider.currentTimeMillis
+            if ((now - currentSegmentTimestamp.time >= options.experimental.sessionReplay.sessionSegmentDuration)) {
                 val segment =
-                    createSegment(
+                    createSegmentInternal(
                         options.experimental.sessionReplay.sessionSegmentDuration,
                         currentSegmentTimestamp,
-                        replayId,
-                        segmentId,
+                        currentReplayId,
+                        currentSegment,
                         height,
                         width
                     )
                 if (segment is ReplaySegment.Created) {
                     segment.capture(hub)
-                    currentSegment.getAndIncrement()
+                    currentSegment++
                     // set next segment timestamp as close to the previous one as possible to avoid gaps
-                    segmentTimestamp.set(DateUtils.getDateTime(currentSegmentTimestamp.time + segment.videoDuration))
+                    segmentTimestamp = DateUtils.getDateTime(currentSegmentTimestamp.time + segment.videoDuration)
                 }
-            } else if ((now - replayStartTimestamp.get() >= options.experimental.sessionReplay.sessionDuration)) {
-                stop()
+            }
+
+            if ((now - replayStartTimestamp.get() >= options.experimental.sessionReplay.sessionDuration)) {
+                options.replayController.stop()
                 options.logger.log(INFO, "Session replay deadline exceeded (1h), stopping recording")
             }
         }
     }
 
     override fun onConfigurationChanged(recorderConfig: ScreenshotRecorderConfig) {
-        val currentSegmentTimestamp = segmentTimestamp.get()
+        val currentSegmentTimestamp = segmentTimestamp ?: return
         createCurrentSegment("onConfigurationChanged") { segment ->
             if (segment is ReplaySegment.Created) {
                 segment.capture(hub)
 
-                currentSegment.getAndIncrement()
+                currentSegment++
                 // set next segment timestamp as close to the previous one as possible to avoid gaps
-                segmentTimestamp.set(DateUtils.getDateTime(currentSegmentTimestamp.time + segment.videoDuration))
+                segmentTimestamp = DateUtils.getDateTime(currentSegmentTimestamp.time + segment.videoDuration)
             }
         }
 
@@ -137,15 +141,15 @@ internal class SessionCaptureStrategy(
 
     private fun createCurrentSegment(taskName: String, onSegmentCreated: (ReplaySegment) -> Unit) {
         val now = dateProvider.currentTimeMillis
-        val currentSegmentTimestamp = segmentTimestamp.get()
-        val segmentId = currentSegment.get()
-        val duration = now - (currentSegmentTimestamp?.time ?: 0)
-        val replayId = currentReplayId.get()
+        val currentSegmentTimestamp = segmentTimestamp ?: return
+        val segmentId = currentSegment
+        val duration = now - currentSegmentTimestamp.time
+        val replayId = currentReplayId
         val height = recorderConfig.recordingHeight
         val width = recorderConfig.recordingWidth
         replayExecutor.submitSafely(options, "$TAG.$taskName") {
             val segment =
-                createSegment(duration, currentSegmentTimestamp, replayId, segmentId, height, width)
+                createSegmentInternal(duration, currentSegmentTimestamp, replayId, segmentId, height, width)
             onSegmentCreated(segment)
         }
     }