+++ sentry-android-replay/src/main/java/io/sentry/android/replay/video/SimpleVideoEncoder.kt
@@ -120,7 +120,7 @@ internal class SimpleVideoEncoder(
         )
         format.setInteger(MediaFormat.KEY_BIT_RATE, bitRate)
         format.setFloat(MediaFormat.KEY_FRAME_RATE, muxerConfig.frameRate.toFloat())
-        format.setInteger(MediaFormat.KEY_I_FRAME_INTERVAL, 10)
+        format.setInteger(MediaFormat.KEY_I_FRAME_INTERVAL, -1) // use -1 to force always non-key frames, meaning only partial updates to save the video size
 
         format
     }